Split the code up into separate files and did a bit of cleanup.
